color_processor/requirements.txt
nats-py
opencv-contrib-python-headless
numpy
webcolors
pydantic
pytest